+++ fptcloud/floating-ip/resource_floating_ip.go
@@ -5,7 +5,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"log"
 	common "terraform-provider-fptcloud/commons"
 	"time"
@@ -15,42 +14,13 @@ import (
 // This can be used to create, read and delete operations for a floating ip in the infrastructure.
 func ResourceFloatingIp() *schema.Resource {
 	return &schema.Resource{
-		Description: "Provides a FPT cloud instance group which can be attached to an instance in order to provide expanded floating ip.",
+		Description: "Provides a FPT cloud floating ip which can be created to public ip address in order to provide expanded floating ip.",
 		Schema: map[string]*schema.Schema{
 			"vpc_id": {
 				Type:        schema.TypeString,
 				Required:    true,
 				Description: "The vpc id of the floating ip",
 				ForceNew:    true,
-			},
-			"floating_ip_id": {
-				Type:         schema.TypeString,
-				Required:     true,
-				ValidateFunc: validation.NoZeroValues,
-				Description:  "The id of the ip address",
-				ForceNew:     true,
-			},
-			"instance_id": {
-				Type:        schema.TypeString,
-				Optional:    true,
-				Description: "The id of the instance",
-				ForceNew:    true,
-			},
-			"floating_ip_port": {
-				Type:         schema.TypeInt,
-				Optional:     true,
-				ValidateFunc: validation.IntBetween(1, 65535),
-				RequiredWith: []string{"instance_id", "instance_port"},
-				Description:  "The port of the ip address",
-				ForceNew:     true,
-			},
-			"instance_port": {
-				Type:         schema.TypeInt,
-				Optional:     true,
-				ValidateFunc: validation.IntBetween(1, 65535),
-				RequiredWith: []string{"instance_id", "floating_ip_port"},
-				Description:  "The port of the instance",
-				ForceNew:     true,
 			},
 			"ip_address": {
 				Type:     schema.TypeString,
@@ -74,31 +44,17 @@ func resourceFloatingIpCreate(ctx context.Context, d *schema.ResourceData, m int
 	apiClient := m.(*common.Client)
 	service := NewFloatingIpService(apiClient)
 
-	createModel := CreateFloatingIpDTO{}
 	vpcId, okVpcId := d.GetOk("vpc_id")
-	if okVpcId {
-		createModel.VpcId = vpcId.(string)
+	if !okVpcId {
+		return diag.Errorf("[ERR] Vpc id is required")
 	}
-	if floatingIpId, ok := d.GetOk("floating_ip_id"); ok {
-		createModel.FloatingIpId = floatingIpId.(string)
-	}
-	if floatingIpPort, ok := d.GetOk("floating_ip_port"); ok {
-		createModel.FloatingIpPort = floatingIpPort.(int)
-	}
-	if instanceId, ok := d.GetOk("instance_id"); ok {
-		createModel.InstanceId = instanceId.(string)
-	}
-	if instancePort, ok := d.GetOk("instance_port"); ok {
-		createModel.InstancePort = instancePort.(int)
-	}
-
-	result, err := service.CreateFloatingIp(createModel)
+	result, err := service.CreateFloatingIp(vpcId.(string))
 	if err != nil || result == nil {
 		return diag.Errorf("[ERR] Failed to create a new floating ip: %s", err)
 	}
 
 	var setError error
-	d.SetId("")
+	d.SetId(result.ID)
 	setError = d.Set("vpc_id", vpcId)
 	setError = d.Set("ip_address", result.IpAddress)
 	if setError != nil {
@@ -107,25 +63,23 @@ func resourceFloatingIpCreate(ctx context.Context, d *schema.ResourceData, m int
 
 	//Waiting for status active
 	createStateConf := &retry.StateChangeConf{
-		Pending: []string{"INACTIVE"},
+		Pending: []string{},
 		Target:  []string{"ACTIVE", "IN_ACTIVE"},
 		Refresh: func() (interface{}, string, error) {
 			findModel := FindFloatingIpDTO{
-				IpAddress: result.IpAddress,
-				VpcId:     vpcId.(string),
+				FloatingIpID: result.ID,
+				VpcId:        vpcId.(string),
 			}
-			resp, err := service.FindFloatingIpByAddress(findModel)
+			resp, err := service.FindFloatingIp(findModel)
 			if err != nil {
 				return 0, "", common.DecodeError(err)
 			}
-			d.SetId(resp.ID)
 			return resp, resp.Status, nil
 		},
-		Timeout:                   5 * time.Minute,
-		Delay:                     30 * time.Second,
-		MinTimeout:                30 * time.Second,
-		ContinuousTargetOccurence: 3,
-		NotFoundChecks:            20,
+		Timeout:        5 * time.Minute,
+		Delay:          3 * time.Second,
+		MinTimeout:     3 * time.Second,
+		NotFoundChecks: 120,
 	}
 	_, err = createStateConf.WaitForStateContext(ctx)
 	if err != nil {
@@ -157,10 +111,8 @@ func resourceFloatingIpRead(_ context.Context, d *schema.ResourceData, m interfa
 
 	var setError error
 	d.SetId(result.ID)
+	setError = d.Set("vpc_id", findModel.VpcId)
 	setError = d.Set("ip_address", result.IpAddress)
-	setError = d.Set("nat_type", result.NatType)
-	setError = d.Set("instance", result.Instance)
-	setError = d.Set("status", result.Status)
 	setError = d.Set("created_at", result.CreatedAt)
 	if setError != nil {
 		return diag.Errorf("[ERR] Floating ip could not be found")