+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [0.1.2] - 2024-09-15
+### Resource
+- Update schema instance (using flavor name and image name)
+- Fix bug security group rule
+
+## [0.1.1] - 2024-08-10
+### Data source
+- Subnet
+### Resource
+- Floating IP association
+- Subnet
+
 ## [0.1.0] - 2024-08-09
 ### Data source
 - Floating IP